+++ test/fixture/local-config/initConfig/plugin2.js
@@ -1,8 +1,8 @@
 module.exports = function() {
   'use strict';
 
-  var orig = this.learn('plugin2.p2k2.d');
-  this.demand('plugin2.p2k2.d', orig.concat('merge'));
+  var orig = this.learn('initConfig.plugin2.p2k2.d');
+  this.demand('initConfig.plugin2.p2k2.d', orig.concat('merge'));
 
   return {
     p2k3: {e: 5, f: 6},